New: Extract JavaScript parser
