+++ packages/parser-javascript/README.md
@@ -1,0 +1,2 @@
+# Parser javascript (`@sonarwhal/parser-javascript`)
+
