+++ packages/sonarwhal/package.json
@@ -16,8 +16,6 @@
     "chrome-remote-interface": "^0.25.5",
     "content-type": "^1.0.4",
     "debug": "^3.1.0",
-    "eslint": "^4.18.1",
-    "espree": "^3.5.3",
     "eventemitter2": "^5.0.1",
     "file-type": "^7.5.0",
     "file-url": "^2.0.2",
@@ -69,6 +67,7 @@
     "async-retry": "^1.1.4",
     "ava": "^0.25.0",
     "cpx": "^1.5.0",
+    "eslint": "^4.18.0",
     "eslint-plugin-import": "^2.8.0",
     "eslint-plugin-markdown": "^1.0.0-beta.7",
     "eslint-plugin-typescript": "^0.9.0",
